requirements.txt
pandas
matplotlib
python-dotenv
openai
litellm
